Create backup command.
The backup command sends and receives full or incrementall snapshots.
The command requires as arguments two ZFS filesystems - source
and destination. The destination may be located on a remote host.
In this case snapshots are transfered via ssh.

If the destination does not exist the last source snapshot is
transfered, otherwise an incremental snapshot between the last
destination snapshot - must exist at the source - and the last source
snapshot is transfered.

The snapshots may be filtered by prefix.